move rigs contract abi to /abis & import contract address from env
